refactor:Cambio de nombre de pagina a HeadquartersManagementPage
